fix:调整页面问题

+++ src/main/java/com/tangerineSpecter/oms/system/domain/vo/data/TradeRecordInfoVo.java
@@ -14,9 +14,9 @@ public class TradeRecordInfoVo {
 
     private Long id;
     @NotNull(message = "开盘资金不能为空")
-    private Integer startMoney;
+    private Double startMoney;
     @NotNull(message = "收盘资金不能为空")
-    private Integer endMoney;
+    private Double endMoney;
     @NotBlank(message = "交易时间不能为空")
     private String date;
     @NotNull(message = "类型不能为空")
